+++ src/test/kotlin/id/walt/vclib/schema/SchemaServiceTest.kt
@@ -2,7 +2,7 @@ package id.walt.vclib.schema
 
 import com.beust.klaxon.Json
 import id.walt.vclib.credentials.*
-import id.walt.vclib.model.VerifiableCredential
+import id.walt.vclib.model.*
 import id.walt.vclib.registry.VerifiableCredentialMetadata
 import id.walt.vclib.schema.SchemaService.DateTimeFormat
 import id.walt.vclib.schema.SchemaService.JsonIgnore
@@ -11,11 +11,20 @@ import io.kotest.core.spec.style.StringSpec
 import java.io.File
 
 data class DummyCredential(
-    @Json(name = "@context") @field:PropertyName(name = "@context") @field:SchemaService.Required var context: List<String> = listOf("https://www.w3.org/2018/credentials/v1"),
+    @Json(name = "@context") @field:PropertyName(name = "@context") @field:SchemaService.Required var context: List<String> = listOf(
+        "https://www.w3.org/2018/credentials/v1"
+    ),
     @Json(serializeNull = false) @field:SchemaService.Required override var id: String? = null,
-    @Json(serializeNull = false) @field:DateTimeFormat var issuanceDate: String? = null,
-    @field:JsonIgnore var toIgnore: String? = null
-) : VerifiableCredential(type) {
+    @Json(serializeNull = false) @field:DateTimeFormat override var issuanceDate: String? = null,
+    @field:JsonIgnore var toIgnore: String? = null,
+    @Json(serializeNull = false) override var issuer: String? = null,
+    @Json(serializeNull = false) override var validFrom: String? = null,
+    @Json(serializeNull = false) override var expirationDate: String? = null,
+    @Json(serializeNull = false) override var credentialSchema: CredentialSchema? = null,
+    @Json(serializeNull = false) override var proof: Proof? = null,
+    @Json(serializeNull = false) override var credentialSubject: DummySubject? = null
+) : AbstractVerifiableCredential<DummyCredential.DummySubject>(type) {
+    data class DummySubject(override var id: String?) : CredentialSubject()
     companion object : VerifiableCredentialMetadata(
         type = listOf("VerifiableCredential", "VerifiableAttestation", "Dummy"),
         template = { DummyCredential() }
@@ -28,56 +37,56 @@ class SchemaServiceTest : StringSpec({
         generateSchema(DummyCredential::class.java)
     }
 
-    "verify VerifiableDiploma schema"   {
+    "testing VerifiableDiploma schema"   {
         generateSchema(VerifiableDiploma::class.java)
     }
 
-    "verify VerifiableAttestation schema"   {
+    "testing VerifiableAttestation schema"   {
         generateSchema(VerifiableAttestation::class.java)
     }
 
-    "verify VerifiableAuthorization schema"   {
+    "testing VerifiableAuthorization schema"   {
         generateSchema(VerifiableAuthorization::class.java)
     }
 
-    "verify VerifiableId schema"   {
+    "testing VerifiableId schema"   {
         generateSchema(VerifiableId::class.java)
     }
 
-    "verify Europass schema"   {
+    "testing Europass schema"   {
         generateSchema(Europass::class.java)
     }
 
-    "verify UniversityDegree schema"   {
+    "testing UniversityDegree schema"   {
         generateSchema(UniversityDegree::class.java)
     }
 
-    "verify PermanentResidentCard schema"   {
+    "testing PermanentResidentCard schema"   {
         generateSchema(PermanentResidentCard::class.java)
     }
 
-    "verify GaiaxCredential schema"   {
+    "testing GaiaxCredential schema"   {
         generateSchema(GaiaxCredential::class.java)
     }
 
-    "verify GaiaxSelfDescription schema"   {
+    "testing GaiaxSelfDescription schema"   {
         generateSchema(GaiaxSelfDescription::class.java)
     }
 
-    "verify GaiaxServiceOffering schema"   {
+    "testing GaiaxServiceOffering schema"   {
         generateSchema(GaiaxServiceOffering::class.java)
     }
 
-    "verify VerifiableVaccinationCertificate schema"   {
+    "testing VerifiableVaccinationCertificate schema"   {
         generateSchema(VerifiableVaccinationCertificate::class.java)
     }
 
-    "verify ProofOfResidence schema"   {
+    "testing ProofOfResidence schema"   {
         generateSchema(ProofOfResidence::class.java)
     }
 
-    "verify VerifiableMandate schema"   {
-        generateSchema(VerifiableMandate::class.java)
+    "testing ParticipantCredential schema"   {
+        generateSchema(ParticipantCredential::class.java)
     }
 })